% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_elevation}
\alias{bci_elevation}
\title{Elevation data from Barro Colorado Island (BCI), Panama.}
\format{A tibble with 20301 rows and 3 variables:
\itemize{
\item \code{x}, \code{y}: coordinates at every corner across the plot (xxx in meters from
some reference).
\item \code{elev}: elevation at every corner across the plot (xxx in meters from some
reference).
}}
\source{
Lao, Suzanne \email{LAOZ@si.edu}.
}
\usage{
bci_elevation
}
\description{
Elevation data from Barro Colorado Island (BCI), Panama.
}
\keyword{datasets}
