% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_habitat}
\alias{bci_habitat}
\title{Habitat data from Barro Colorado Island (BCI), Panama.}
\format{A tibble with 1250 rows and 3 variables:
\itemize{
\item \code{x}, \code{y}: coordinates at every corner across the plot (xxx in meters from
some reference).
\item \code{habitat}: Habitats per 20x20m quadrats determined by observation.
}}
\source{
Lao, Suzanne \email{LAOZ@si.edu}.
}
\usage{
bci_habitat
}
\description{
Habitats from BCI per 20x20m quadrats determined by observation. Habitats can
also be determined by the slope, convexity, and elevation.
}
\keyword{datasets}
