% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_mat}
\alias{bci_mat}
\title{xxx Ask Suzanne.}
\format{A matrix with 101 rows and 201 columns. Row and column names are NULL.}
\source{
Lao, Suzanne \email{LAOZ@si.edu}.
}
\usage{
bci_mat
}
\description{
xxx Ask Suzanne.
}
\keyword{datasets}
