% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_plotdim}
\alias{bci_plotdim}
\title{Dimensions of the plot from Barro Colorado Island (BCI).}
\format{An unnamed vector of length 2.}
\source{
Lao, Suzanne \email{LAOZ@si.edu}.
}
\usage{
bci_plotdim
}
\description{
Dimensions of the plot from Barro Colorado Island (BCI).
}
\keyword{datasets}
