% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_quad_info}
\alias{bci_quad_info}
\title{Habitat data from Barro Colorado Island (BCI), Panama.}
\format{A tibble with 1250 rows and 5 variables xxx:
\itemize{
\item \code{elev}: Elevation.
\item \code{convex}: Convexity.
\item \code{slope}: Slope.
\item \code{hab}: Habitat. xxx I converted this from factor to integer.
\item \code{quad}: xxx I converted this from factor to integer
}}
\source{
Lao, Suzanne \email{LAOZ@si.edu}.
}
\usage{
bci_quad_info
}
\description{
Habitats from BCI determined by the slope, convexity, and elevation.
}
\keyword{datasets}
