% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_species}
\alias{bci_species}
\title{Species from Barro Colorado Island (BCI), Panama.}
\format{A tibble with 1414 rows and 13 variables.
Variable names at http://ctfs.si.edu/Public/DataDict/data_dict.php.
\itemize{
\item \code{sp}: Species code.
\item \code{Latin}: Latin name.
\item \code{Genus}: Genus name.
\item \code{Species}: Species part of Latin name; (or may be a morphospecies name).
\item \code{Family}: Family name.
\item \code{SpeciesID}: Primary key, an integer automatically generated to uniquely
identify a    taxonomic species.
\item \code{SubspeciesID}:  An integer automatically generated to uniquely identify a
subspecies.
\item \code{Authority}: Taxonomic authority for the classification of the species.
\item \code{IDLevel}: The deepest taxonomic level for which full identification is
known. Limited to values species, genus, family, none, or multiple. None is
used when family is not known. Multiple is used when the name may include a
mixture of more than one species.
\item \code{syn}: xxx
\item \code{subsp}: Subspecies portion of the Latin name, may be a subspecies or
variety.
\item \code{wsg}: xxx
\item \code{wsglevel}: xxx
}}
\source{
Lao, Suzanne \email{LAOZ@si.edu}.
}
\usage{
bci_species
}
\description{
Species from Barro Colorado Island (BCI), Panama.
}
\keyword{datasets}
