% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_wood_density}
\alias{bci_wood_density}
\title{Wood density data from Barro Colorado Island (BCI), Panama.}
\format{A tibble with 16558 rows and 9 variables xxx:
\itemize{
\item \code{wsg}:
\item \code{idlevel}:
\item \code{site}:
\item \code{sp}:
\item \code{genus}:
\item \code{species}:
\item \code{genwood}:
\item \code{famwood}:
\item \code{spwood}:
}}
\source{
Lao, Suzanne \email{LAOZ@si.edu}.
}
\usage{
bci_wood_density
}
\description{
Wood density data from Barro Colorado Island (BCI), Panama.
}
\keyword{datasets}
