% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bciex-package.r
\docType{package}
\name{bciex}
\alias{bciex}
\alias{bciex-package}
\title{Example Data of Forest Dynamics from Barro Colorado Island (BCI).}
\description{
Example Data of Forest Dynamics from Barro Colorado Island (BCI). The goal of
these data is to make it easy to run examples and tests. But they are
incomplete so you should not used it for analyses.
}
