% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bciYYsNmini}
\alias{bciYYsNmini}
\alias{bci12s1mini}
\alias{bci12s2mini}
\alias{bci12s3mini}
\alias{bci12s4mini}
\alias{bci12s5mini}
\alias{bci12s6mini}
\alias{bci12s7mini}
\title{A randomly selected subset of stem tables from Barro Colorado Island, Panama.}
\format{A dataframe with 2,165 rows and 20 variables.}
\source{
Files bci.stem#.rdata from
\url{https://repository.si.edu/handle/10088/20925}.
}
\description{
For every bciYYsNmini "YY" and "N" indicate the year and census number. For
example, bci12s1mini contains data released in 2012 from census 1.
}
\seealso{
census

Other census tables.: \code{\link{bciYYtNmini}},
  \code{\link{census}}
}
