% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bciYYtNmini}
\alias{bciYYtNmini}
\alias{bci12t1mini}
\alias{bci12t2mini}
\alias{bci12t3mini}
\alias{bci12t4mini}
\alias{bci12t5mini}
\alias{bci12t6mini}
\alias{bci12t7mini}
\title{A randomly selected subset of tree tables from Barro Colorado Island, Panama.}
\format{A dataframe with 1,000 rows and 20 variables.}
\source{
Files bci.full#.rdata from
\url{https://repository.si.edu/handle/10088/20925}.
}
\description{
For every bciYYsNmini "YY" and "N" indicate the year and census number. For
example, bci12t1mini contains data released in 2012 from census 1.
}
\seealso{
census

Other census tables.: \code{\link{bciYYsNmini}},
  \code{\link{census}}
}
