% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bciex-package.r
\docType{package}
\name{bciex}
\alias{bciex}
\alias{bciex-package}
\title{Forest Dynamics Data from Barro Colorado Island (BCI), 2016.}
\description{
Forest Dynamics Data from Barro Colorado Island (BCI), 2016.
}
